requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.5.2
Babel==2.3.4
backports.functools-lru-cache==1.4
cffi==1.10.0
cliff==2.7.0
cmd2==0.7.0
configparser==3.5.0
cryptography==1.8.2
debtcollector==1.13.0
deprecation==1.0.1
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
iso8601==0.1.11
isort==4.2.5
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.20.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
openstacksdk==0.9.16
os-client-config==1.27.0
osc-lib==1.6.0
oslo.config==4.1.1
oslo.i18n==3.15.1
oslo.serialization==2.18.0
oslo.utils==3.25.1
packaging==16.8
paramiko==2.1.2
pbr==3.0.1
# pkg-resources==0.0.0
positional==1.1.1
prettytable==0.7.2
pyasn1==0.2.3
pycparser==2.17
pylint==1.7.1
pyparsing==2.2.0
python-cinderclient==2.0.1
python-glanceclient==2.6.0
python-keystoneclient==3.10.0
python-novaclient==8.0.0
python-openstackclient==3.11.0
pytz==2017.2
PyYAML==3.12
requests==2.14.2
requestsexceptions==1.2.0
rfc3986==1.0.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.21.0
unicodecsv==0.14.1
warlock==1.2.0
wrapt==1.10.10
